requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.2
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
